[#2] Refactor: mock data 가공 모듈 가독성 좋게 세분화
