+++ src/utils/index.ts
@@ -1,2 +1,2 @@
 export * from './fetch';
-export * from './processChartData';
+export * from './chartHelpers';
